requirements_complete.txt
# Codex Dominion Complete Platform Requirements
# 🔥 Digital Sovereignty Dependencies 👑

# Core Framework
streamlit>=1.28.0
streamlit-authenticator>=0.2.3

# Data Processing & Analytics
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Social Media Integration
tweepy>=4.14.0
requests>=2.31.0
urllib3>=2.0.0
python-buffer>=0.2.0

# E-commerce Integration
woocommerce>=3.0.0
requests-oauthlib>=1.3.1

# API & Web Services
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
aiohttp>=3.8.0

# Data Storage & Caching
redis>=4.5.0
sqlite3
python-dotenv>=1.0.0

# Date & Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# Image & Media Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Security & Authentication
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0

# Configuration & Logging
pyyaml>=6.0.0
python-json-logger>=2.0.0
configparser>=5.3.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Cloud & Deployment
google-cloud-storage>=2.10.0
google-cloud-run>=0.10.0
boto3>=1.28.0

# Monitoring & Performance
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional Utilities
schedule>=1.2.0
python-slugify>=8.0.0
emoji>=2.8.0
python-dateutil>=2.8.2